requirements.txt
pytest~=6.2.4
requests~=2.25.1
python-dotenv~=0.17.1
PyJWT~=2.1.0
psycopg2~=2.8.6
python-dateutil~=2.8.1